requirements.txt
discord.py[voice]
discord.py
requests
yt-dlp
PyNaCl
ffmpeg-python
python-dotenv